requirements.txt
Django==4.1.7
djangorestframework==3.14.0
markdown==3.4.3
django-filter==23.1
django-autoslug==1.9.8
python-dotenv==1.0.0
Pillow==9.4.0
django-cors-headers==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
gunicorn==21.2.0
whitenoise==6.5.0